refactor(warnings): split visual evidence helpers
Separate text-layer, graphics, annotation, and shared geometry helpers so each warning module has a narrower responsibility while preserving the public visualEvidence import surface.
