+++ src/core/warnings/visualEvidence/textComparison.ts
@@ -1,0 +1,20 @@
+export function normalizeComparableText(text: string): string {
+  return text
+    .normalize('NFKC')
+    .toLowerCase()
+    .replace(/[^\p{L}\p{N}]+/gu, '');
+}
+
+export function levenshteinDistance(a: string, b: string): number {
+  const previous = Array.from({ length: b.length + 1 }, (_, index) => index);
+  const current = Array.from({ length: b.length + 1 }, () => 0);
+  for (let i = 0; i < a.length; i++) {
+    current[0] = i + 1;
+    for (let j = 0; j < b.length; j++) {
+      const substitution = previous[j] + (a[i] === b[j] ? 0 : 1);
+      current[j + 1] = Math.min(previous[j + 1] + 1, current[j] + 1, substitution);
+    }
+    for (let j = 0; j < previous.length; j++) previous[j] = current[j];
+  }
+  return previous[b.length] ?? 0;
+}
